XMLHttpRequest/abort-after-timeout.htm

<!doctype html>
<html>
<head>
    <title>XMLHttpRequest: abort() after successful receive should not fire "abort" event</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-abort()-method" />
</head>
<body>
<div id="log"></div>
<script>
    var test = async_test();

    test.step(function() {
        // timeout is 100ms
        // the download would otherwise take 500ms
        // we check after 300ms to make sure abort does not fire an "abort" event

        var client = new XMLHttpRequest();
        client.timeout = 100;

        setTimeout(test.step_func(function() {
            // abort should not cause the "abort" event to fire

            client.abort();

            assert_equals(client.readyState, 0);

            test.done();
        }), 300);

        client.onabort = test.step_func(function () {
            // this should not fire!

            assert_unreached("abort() should not cause the abort event to fire");
        });

        client.open("GET", "resources/delay.php?ms=500000", true);
        client.send(null);
    });
</script>
</body>
</html>
